ipam/host-local: Move allocator and config to backend
